Preserve input dtype in spatial_tracker_v2 positional embeddings
make_sincos_pos_embed forced float32 output, promoting bfloat16 tokens to
float32 before the DPT head's ConvTranspose2d layers — which carry bfloat16
weights when loaded with dtype_override — and failing with a dtype mismatch.

+++ spatial_tracker_v2/pytorch/src/heads/utils.py
@@ -39,7 +39,7 @@ def position_grid_to_embed(
     # Combine and reshape
     emb = torch.cat([emb_x, emb_y], dim=-1)  # [1, H*W, D]
 
-    return emb.view(H, W, embed_dim)  # [H, W, D]
+    return emb.view(H, W, embed_dim).to(pos_grid.dtype)  # [H, W, D]
 
 
 def make_sincos_pos_embed(
